Fix: keyboard v2?

+++ src/header/text/framebuffer.h
@@ -6,7 +6,7 @@
 #include <stddef.h>
 
 // Before: 0xB8000
-#define FRAMEBUFFER_MEMORY_OFFSET ((uint32_t*) 0xC00B8000) 
+#define FRAMEBUFFER_MEMORY_OFFSET ((uint64_t*) 0xC00B8000) 
 #define CURSOR_PORT_CMD    0x03D4
 #define CURSOR_PORT_DATA   0x03D5
 #define WIDTH 80
